Move swap phase 2 tests into a seperate file. Remove unused imports.
